Se genero el codigo paara que se pueda guardar los productos manual mente y los mandamos a guaradar a su respectivo archivo de texto json  con su clase persistencia

+++ Producto.cs
@@ -13,13 +13,13 @@ namespace proyecto1programacion
         string marca;
         decimal precio_compra;
         decimal precio_venta;
-        int cantidad_existente;
+        decimal cantidad_existente;
 
         public string Codigo { get => codigo; set => codigo = value; }
         public string Nombre { get => nombre; set => nombre = value; }
         public string Marca { get => marca; set => marca = value; }
         public decimal Precio_compra { get => precio_compra; set => precio_compra = value; }
         public decimal Precio_venta { get => precio_venta; set => precio_venta = value; }
-        public int Cantidad_existente { get => cantidad_existente; set => cantidad_existente = value; }
+        public decimal Cantidad_existente { get => cantidad_existente; set => cantidad_existente = value; }
     }
 }
